backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.0
pydantic-settings==2.6.1
pydantic[email]==2.10.2
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-multipart==0.0.18
pdfplumber==0.11.4
anthropic==0.39.0
openai==1.57.0
python-dateutil==2.9.0
